hello triangle running

+++ gengar_engine/src/engine/vectors.rs
@@ -13,17 +13,17 @@ impl VecThreeFloat {
 // Used for any FFI C stuff. Like ogl graphics rendering
 #[repr(C)]
 pub struct VecThreeFloatC {
-    pub x: f64,
-    pub y: f64,
-    pub z: f64,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl From<VecThreeFloat> for VecThreeFloatC {
     fn from(input: VecThreeFloat) -> Self {
         VecThreeFloatC {
-            x: input.x,
-            y: input.y,
-            z: input.z,
+            x: input.x as f32,
+            y: input.y as f32,
+            z: input.z as f32,
         }
     }
 }
